chore(run-protocol): move attestation out of zoe package
