+++ packages/run-protocol/exported.js
@@ -1,2 +1,3 @@
+import './src/attestation/types.js';
 import './src/vaultFactory/types.js';
 import './src/vpool-xyk-amm/types.js';
