+++ packages/run-protocol/test/attestation/returnableNFT/test-userFlow.js
@@ -5,9 +5,9 @@ import { test } from '@agoric/zoe/tools/prepare-test-env-ava.js';
 import { AmountMath, makeIssuerKit } from '@agoric/ertp';
 import { E } from '@agoric/eventual-send';
 
-import { setupZCFTest } from '../../../zcf/setupZcfTest.js';
+import { setupZCFTest } from '@agoric/zoe/test/unitTests/zcf/setupZcfTest.js';
 
-import { setupAttestation } from '../../../../../src/contracts/attestation/returnable/returnableNFT.js';
+import { setupAttestation } from '../../../src/attestation/returnable/returnableNFT.js';
 
 const makeDoReturnAttestation =
   (zoe, issuer, makeReturnAttInvitation) => async attestation => {
