+++ packages/zoe/src/contracts/exported.js
@@ -2,4 +2,3 @@ import './types.js';
 import './loan/types.js';
 import './priceAggregatorTypes.js';
 import './callSpread/types.js';
-import './attestation/types.js';
